apps/py-loader/requirements.txt
neo4j
fastapi
uvicorn
streamlit
python-dotenv
sentence_transformers